requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
mysqlclient==2.0.3
pkg_resources==0.0.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
